feat: complete refactor data faceted table
