+++ apps/app/app/(wallet)/[address]/overview/nfts/(components)/data-table/data-table-toolbar.tsx
@@ -21,8 +21,8 @@ import { useQueryClient } from "@tanstack/react-query";
 import type { Table } from "@tanstack/react-table";
 import { useMemo } from "react";
 import type { Address } from "viem";
-import { DataTableFacetedFilter } from "@/app/(wallet)/[address]/overview/nfts/(components)/data-table/data-table-faceted-filter";
 import { DataTableViewOptions } from "@/app/(wallet)/[address]/overview/nfts/(components)/data-table/data-table-view-options";
+import { DataTableFacetedFilter } from "@/components/data-table/data-table-faceted-filter";
 import type { NftData, NftDataPage, WalletSettingsData } from "@/data";
 import { queries } from "@/queries";
 import { useAuth } from "@/stores/useAuth";
